Folder_Tavily/requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
tavily-python>=0.7.23
